Fix ECMAScript character class escapes, uint overflow in defineProperties, Temporal timezone
- JsRegExp: replace .NET \s\S\w\W\d\D with ECMAScript-accurate character classes (24 tests)
- JsObject: uint→long cast fix for array index 4294967294
- TemporalHelper: add CanonicalizeTimeZoneId
- Deleted Object_defineProperties and RegExp_CharacterClassEscapes todo files

+++ src/Asynkron.JsEngine/JsTypes/JsObject.cs
@@ -2349,7 +2349,7 @@ public sealed class JsObject : IDictionary<string, object?>, IJsObjectLike,
         numericKeys.Sort();
         foreach (var index in numericKeys)
         {
-            yield return JsValueCache.GetIndexString((int)index);
+            yield return JsValueCache.GetIndexString((long)index);
         }
 
         foreach (var key in stringKeys)
